Set functions inside of components.
Making functions concerned with specific components sit inside their components. Refactoring for testing.
