+++ src/client/components/App.jsx
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
-import { handleSelectedStation, handlePlayState, seekStation, generateRandomStationId, loadStations, findColor, scrollListener, setStateSelectedStation, muteAudio, lastClickedVolume } from '../util/ClientFunctions.jsx';
 
+import { shuffle } from '../util/ClientFunctions.jsx';
 
 import AudioPlayer from './AudioPlayer.jsx';
 import StationList from './StationList.jsx';
@@ -27,14 +27,113 @@ class App extends Component {
       }
     },
 
-    this.handlePlayState = handlePlayState.bind(this);
-    this.handleSelectedStation = handleSelectedStation.bind(this);
-    this.seekStation = seekStation.bind(this);
-    this.generateRandomStationId = generateRandomStationId.bind(this);
-    this.loadStations = loadStations.bind(this);
-    this.scrollListener = scrollListener.bind(this);
-    this.findColor = findColor.bind(this);
-    this.setStateSelectedStation = setStateSelectedStation.bind(this);
+    this.handlePlayState = this.handlePlayState.bind(this);
+    this.handleSelectedStation = this.handleSelectedStation.bind(this);
+    this.seekStation = this.seekStation.bind(this);
+    this.generateRandomStationId = this.generateRandomStationId.bind(this);
+    this.loadStations = this.loadStations.bind(this);
+    this.scrollListener = this.scrollListener.bind(this);
+    this.findColor = this.findColor.bind(this);
+    this.setStateSelectedStation = this.setStateSelectedStation.bind(this);
+  }
+
+  // Initial API request to build up station collection object.
+  loadStations() {
+    fetch('/api/stations', { mode: 'no-cors' })
+      .then((response) => {
+        return response.json()
+      }).then((json) => {
+        this.setState({ stations: shuffle(json) })
+        console.log('parsed json', json)
+      }).catch((ex) => {
+        console.log('parsing failed', ex)
+      });
+  }
+
+  // Takes a selected station and sets it in app level state, plays it.
+  handleSelectedStation(details) {
+    let station = {
+      id: details.id,
+      name: details.name,
+      stream: details.stream,
+      type: details.type
+    };
+    let test = { key: 'value', tester: 'test' };
+    localStorage.setItem('key', JSON.stringify(station)); //sets the localStorage to the station before setting the this.State to the station
+    this.setState({
+      selectedStation: station
+    });
+    this.handlePlayState();
+  }
+
+  handlePlayState() {
+    let player = document.getElementById("player");
+    document.getElementById("player").load();
+
+    // Displays the loading icon while the media is loading.
+    player.addEventListener("loadstart", function () {
+      pButton.className = "fa fa-spinner fa-pulse fa-2x fa-fw";
+    });
+
+    // Plays the audio when it is ready.
+    player.addEventListener("canplaythrough", function () {
+      document.getElementById("player").play();
+    });
+
+    // Changes the loading icon to a pause icon.
+    player.addEventListener("playing", function () {
+      pButton.className = "";
+      pButton.className = "fa fa-pause fa-2x";
+    });
+  }
+
+  // Helper function for seek functionality.
+  generateRandomStationId() {
+    console.log('gen random called');
+    const seekLength = this.state.stations.length - 1;
+    return Math.floor(Math.random() * seekLength);
+  }
+
+  // Play random station.
+  seekStation() {
+    console.log('seekStation called');
+    const randomStationId = this.generateRandomStationId();
+
+    for (const station of this.state.stations) {
+
+      // Checks to see if the seek matches the currently playing station.
+      if (randomStationId === this.state.selectedStation.id) {
+        this.seekStation();
+        return;
+
+        // Plays the seeked station.
+      } else if (randomStationId === station.id) {
+        this.handleSelectedStation(station);
+      }
+    }
+  }
+
+
+
+  findColor() {
+    const colorIndex = Math.floor(this.state.scrollPercent * this.colors.length);
+    return this.colors[colorIndex];
+  }
+
+  scrollListener() {
+    const scrollListen = window.addEventListener('scroll', () => {
+      const percent = window.scrollY / (document.body.scrollHeight - window.innerHeight);
+      this.setState({ scrollPercent: percent });
+    });
+  }
+
+  // Detects last listened station from users local storage, sets it in app state.
+  setStateSelectedStation() {
+    if (JSON.parse(localStorage.getItem('key'))) {
+      this.setState({
+        selectedStation: JSON.parse(localStorage.getItem('key'))
+      });
+    }
   }
 
   componentDidMount() {
